Generate 7-segments from SVG and add argparsing for generate
My original approach (code not uploaded) was to generate the pixel values for the segments by hand. This approach was really brittle and although it worked, it was non-ideal.

I played around with SVG and realized that if I just take a single 7-segment display digit written in SVG polygons, I could much more easily "rasterize" the SVG and support any image output resolution.

An additional benefit of this is that the original SVG is just a text file, so we can version control it. Then if we eventually want to support the elongated 7-segment display, we can just modify that SVG instead of modifying the rasterizing code.

I added an argparser as well to generate so that we can play around with on/off colors, image output height, etc.

The next step here is to generate all combinations, add the ON segments together, and mask the original template image with the new on segments image.