Generate all segment combinations
My original code (not committed) generated separate image files for each digit. I wanted to avoid having a huge directory of different combinations, so I just saved all the images into a single grid image which can be indexed into.

In the algorithm code, we can just load the grid image and crop it to reverse into individual images. Technically we don't even need to do that and we can just iterate through the large grid image and find the corresponding tile that matches the best.