Add tiling algorithm code
The simplest tiling algorithm calculates the brightnesses of all of the candidate tiles, and all of the target tiles. It then chooses the nearest neighbor tile based on brightness.

I am able to get an output image which definitely has some of the original images structure. However a main issue is if the tile images size is much lower than the viewbox (due to integer rounding I believe)

I was also having some issues choosing the best tile but this was because I wasn't scaling the brightnesses of the candidate tiles accordingly. The max brightness value should be 255 and the min brightness should be 0, so we need to linearly scale these to accomodate that range. Otherwise the difference between each candidate tile is too small.