Add video parsing code from OpenCV example
This is sample code to capture video from some sort of path (webcam, file, etc.) and then output the mosaic next to video

It works! However the webcam resolution is a bit too small to really see the resulting image. To solve this, we need to support much smaller tiles.